Refined Search Feature + Alphabetical Countries
Added search feature. The search event handler now refers to the newly created "regionFilter" state to only search for countries within that filtered set of countries as opposed to props.countries (which searches all countries).

Also sorted the countries alphabetically as soon as the API returned the country list so that all countries will be displayed in order on the results component