Fixed Link Route Issues
Link from a current country ...countryresult/AL was working from main page; however, when accessing a border country and changing only the id value in the URL, the component would not unmount and re render. I fixed this by adding a state for the country ID, and when the button is clicked to access a border country, the state changes. The useeffect listens for this statechange and sends another api request to update the country being displayed on the page. Helpful link here

https://stackoverflow.com/questions/62836374/react-router-does-not-update-component-if-url-parameter-changes